pyproject.toml
[project]
name = "flower-classifier"
version = "0.1.0"
description = "Flower classification using deep learning with PyTorch Lightning"
authors = [
    {name = "Alen Ospanov", email = "ospanovalen04@gmail.com"}
]
readme = "README.md"
[tool.poetry.dependencies]
# python = "^3.10"
python = ">=3.10,<3.14"
python-dotenv = "^1.0.1"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"
dvc = "^3.50.0"
pre-commit = ">=4.2.0,<5.0.0"
dvc-s3 = "^3.1.0"
numpy = "<2.0"
ipykernel = "^6.29.5"
mlflow = "^2.22.1"
timm = "^1.0.15"
pillow = "^11.2.1"
click = "^8.2.1"

[tool.poetry.group.dev.dependencies]
pandas = "^2.2.0"
lightning = "^2.1.3"
matplotlib = "^3.8.4"
torchmetrics = "^1.4.0.post0"
flake8 = "^7.2.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"





[tool.poetry.group.gpu]
optional = true

[tool.poetry.group.gpu.dependencies]
torch = "2.7.1"





[tool.poetry.group.lint]
optional = true

[tool.poetry.group.lint.dependencies]
black = "^24.2.0"
isort = "^5.13.2"
pylint = "^3.2.5"





[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
coverage = "^7.5.4"
pytest = "^8.0.2"
pytest-cov = "^5.0.0"
